EditProduct,Details,DeleteProduct functionality added , still to fix Foreign Key constraint violation between Product and Category

+++ ProductsDistribution/Controllers/CategoryController.cs
@@ -16,13 +16,14 @@ namespace ProductsDistribution.Controllers
     public class CategoryController : Controller
     {
         private readonly ICategoryService categoryService;
-      
+        private readonly IProductService productService;
 
         // GET: Category
 
-        public CategoryController(ICategoryService categoryService)
+        public CategoryController(ICategoryService categoryService,IProductService productService)
         {
             this.categoryService = categoryService;
+            this.productService = productService;
         }
         public ActionResult Index()
         {
@@ -67,17 +68,10 @@ namespace ProductsDistribution.Controllers
             inputModel.categories = GetCategories();
             if (!this.ModelState.IsValid)
             {
-                /* Response.StatusCode = 400; // Replace .AddHeader
-                 var error = new
-                 {
-                     Message = "Request is not valid!"
-                 };
-                 return Json(error);*/
-             //   inputModel.categories = GetCategories();
+               
                 return View(inputModel);
             }
-          //  inputModel.categories = GetCategories();
-
+  
             string selected_category = inputModel.selectedCategory;
             try
             {
@@ -138,25 +132,75 @@ namespace ProductsDistribution.Controllers
             return View(viewModel);
         }
 
-      
+
         [HttpGet]
         public ActionResult Delete(int id)
         {
-            CategoryDTO category = this.categoryService.GetById(id);
-            //vsichki kategorii koito sa deca na category
-            List<CategoryDTO> all_children_for_category = this.categoryService.GetAllSubCategoriesById(id);
-            foreach(CategoryDTO cat in all_children_for_category)
+            if (!ModelState.IsValid)
             {
-                cat.CategoryDTO_parent_id = null;
-                this.categoryService.DeleteCategory(cat);
+
+                return View("DisplayCategories");
             }
-         
-            this.categoryService.DeleteCategory(category);
+            CategoryDTO category = this.categoryService.GetById(id);
+            List<CategoryDTO> all_children_for_category = this.categoryService.GetAllSubCategoriesById(id);
+            if (category == null)
+            {
+                return HttpNotFound();
+            }
+
+           bool isInProducts = this.productService.isInProducts(id);
+
+            bool isParent = this.categoryService.isParent(id);
+
+            try
+            {
+                if (!isInProducts)
+                {
+                    //your deletetion code
+
+                    //    bool isInProducts = this.productService.isInProducts(id);           
+                    foreach (CategoryDTO cat in all_children_for_category)
+                    {
+                        cat.CategoryDTO_parent_id = null;
+                        this.categoryService.DeleteCategory(cat);
+                    }
+
+                    this.categoryService.DeleteCategory(category);
+                }
+            }
+            catch (SqlException ex)
             
 
-            return Redirect(Request.UrlReferrer.ToString());
-        }
+                /*if (ex.Errors.Count > 0) // Assume the interesting stuff is in the first error
+                {
+                    switch (ex.Errors[0].Number)
+                    {
+                        case 547: // Foreign Key violation
+                            ModelState.AddModelError("category_name", "Категорията има прикрепени продукти.За да изтриете категория не трябва да има прикрепени продукти към нея");
+                            return View("DisplayCategories");
+                            //break;
+                        default:
+                            throw;
 
+                    }
+                }*/
+                when (ex.InnerException?.InnerException is SqlException sqlEx &&
+                    (sqlEx.Number == 547))
+                    {
+               
+                ModelState.AddModelError("Категорията има прикрепени продукти.За да изтриете категорията не трябва да има прикрепени продукти към нея",ex.Message);
+                    return View("DisplayCategories");
+                    
+            
+                
+            }
+            // bool isInProducts = this.productService.isInProducts(category.category_id);
+
+            //vsichki kategorii koito sa deca na category
+        
+                return Redirect(Request.UrlReferrer.ToString());
+  
+        }
         CategoryInputEditModel MapCategoryDTOToCategoryInputEditModel(CategoryDTO category)
         {
             return new CategoryInputEditModel()
@@ -173,7 +217,11 @@ namespace ProductsDistribution.Controllers
         public ActionResult EditView(int id)
         {
             var category = this.categoryService.GetById(id);
-         
+
+            if (category == null)
+            {
+                return HttpNotFound();
+            }
             CategoryInputEditModel model = MapCategoryDTOToCategoryInputEditModel(category);
           
             return View(model);
